+++ stories/Schema/ArrayFieldSchema/index.js
@@ -57,10 +57,10 @@ const schema = {
             type: 'string',
             title: 'Are you married?',
             enum: ['yes', 'no'],
-            'ui:control': 'radio',
-            'informed:props': {
-              notify: ['spouse']
-            }
+            'ui:control': 'radio'
+            // 'informed:props': {
+            //   notify: ['spouse']
+            // }
           },
           spouse: {
             type: 'string',