+++ stories/Schema/ArrayFieldSchemaNested/index.js
@@ -58,10 +58,10 @@ const schema = {
             type: 'string',
             title: 'Are you married?',
             enum: ['yes', 'no'],
-            'ui:control': 'radio',
-            'informed:props': {
-              notify: ['spouse']
-            }
+            'ui:control': 'radio'
+            // 'informed:props': {
+            //   notify: ['spouse']
+            // }
           },
           spouse: {
             type: 'string',
@@ -93,10 +93,10 @@ const schema = {
                   type: 'string',
                   title: 'Married?',
                   enum: ['yes', 'no'],
-                  'ui:control': 'radio',
-                  'informed:props': {
-                    notify: ['spouse']
-                  }
+                  'ui:control': 'radio'
+                  // 'informed:props': {
+                  //   notify: ['spouse']
+                  // }
                 },
                 spouse: {
                   type: 'string',